feat: Profile Navigation
